requirements.txt
wheel
Django>=4.0.1
celery>=5.2.3
channels>=3.0.4
channels-redis>=3.3.1
django-celery-beat>=2.2.1
django-celery-results>=2.2.0
django-ninja>=0.13.2
django-extensions>=3.1.5
django-cors-headers>=3.11.0
django-debug-toolbar>=3.2.4
gunicorn>=20.1.0
uvicorn>=0.17.0
pandas>=1.3.5
psycopg2>=2.9.3
pylint>=2.12.2
orjson>=3.6.5
redis>=4.1.0
ipython>=8.0.0
flower==1.0.0